Removed a few make install options which were causing builds from 21 to faile

+++ build.sh
@@ -14,7 +14,7 @@ module add mpc/1.0.1
 echo ${LD_LIBRARY_PATH}
 echo ${PATH}
 
-echo ${MPC_DIR} ${MPFR_DIR} ${ncurses_DIR}
+echo ${MPC_DIR} ${MPC_DIR} ${MPFR_DIR} ${ncurses_DIR}
 
 echo "REPO_DIR is "
 echo ${REPO_DIR}
@@ -31,15 +31,23 @@ mkdir -p ${SOFT_DIR}
 
 #  Download the source file
 
-if [[ ! -s ${SRC_DIR}/${SOURCE_FILE} ]] ; then
+if [ ! -e ${SRC_DIR}/${SOURCE_FILE}.lock ] && [ ! -s ${SRC_DIR}/${SOURCE_FILE} ] ; then
+  touch  ${SRC_DIR}/${SOURCE_FILE}.lock
   echo "seems like this is the first build - let's get the source"
   mkdir -p ${SRC_DIR}
   wget http://mirror.ufs.ac.za/gnu/gnu/${NAME}/${NAME}-${VERSION}/${SOURCE_FILE} -O ${SRC_DIR}/${SOURCE_FILE}
-  tar xzf ${SRC_DIR}/${SOURCE_FILE} -C ${WORKSPACE} --skip-old-files
+  rm -v ${SRC_DIR}/${SOURCE_FILE}.lock
+elif [ -e ${SRC_DIR}/${SOURCE_FILE}.lock ] ; then
+  # Someone else has the file, wait till it's released
+  while [ -e ${SRC_DIR}/${SOURCE_FILE}.lock ] ; do
+    echo " There seems to be a download currently under way, will check again in 5 sec"
+    sleep 5
+  done
 else
   echo "continuing from previous builds, using source at " ${SRC_DIR}/${SOURCE_FILE}
-  tar xzf ${SRC_DIR}/${SOURCE_FILE} -C ${WORKSPACE} --skip-old-files
 fi
+tar xzf ${SRC_DIR}/${SOURCE_FILE} -C ${WORKSPACE} --skip-old-files
+
 cd ${WORKSPACE}/${NAME}-${VERSION}
 
 ./configure --prefix=${SOFT_DIR} \
